# modulo_2_SI/2.1/prac2_bdd/Java/src/main/java/com/davidrey/practica/dtos/requests/BookRequestDto.java
package com.davidrey.practica.dtos.requests;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import javax.validation.constraints.NotBlank;
import java.util.List;


@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class BookRequestDto {

    @NotBlank(message = "Title is mandatory")
    private String title;
    private String summary;
    @NotBlank(message = "Author is mandatory")
    private String author;
    @NotBlank(message = "Publisher is mandatory")
    private String publisher;
    private int publicationYear;
    private List<CommentRequestDto> comments;

}