# modulo_2_SI/2.1/prac2_bdd/Java/src/main/java/com/davidrey/practica/dtos/responses/BookResponseDto.java
package com.davidrey.practica.dtos.responses;

import com.davidrey.practica.models.Comment;
import lombok.Data;

import java.util.List;

@Data
public class BookResponseDto {

    private Long id;
    private String title;
    private List<Comment> comments;

}