# modulo_2_SI/2.1/prac2_bdd/Java/src/main/java/com/davidrey/practica/models/Comment.java
package com.davidrey.practica.models;

import com.fasterxml.jackson.annotation.JsonIgnore;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import javax.persistence.*;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Entity
public class Comment {

    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private Long id;
    private String comment;
    private float score;
    @ManyToOne
    @JsonIgnore
    private Book book;
    @ManyToOne
    private User user;


}